+++ lib/src/features/menu/widgets/cart_quantity_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
+import '../../../constants.dart';
 
 class CartQuantityWidget extends StatelessWidget {
   final String itemName;
