+++ lib/src/features/menu/widgets/category_button_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
+import '../../../constants.dart';
 
 class CategoryButtonWidget extends StatelessWidget {
   final String category;
