+++ lib/src/features/menu/widgets/category_section_widget.dart
